+++ lib/models/roleplay/arda-date/enums.d.ts
@@ -1,0 +1,16 @@
+export declare enum ArdaDateAge {
+    DaysBeforeDays = "daysBeforeDays",
+    YearsOfTheLamps = "yearsOfTheLamps",
+    YearsOfTheTrees = "yearsOfTheTrees",
+    FirstAge = "firstAge",
+    SecondAge = "secondAge",
+    ThirdAge = "thirdAge"
+}
+export declare enum ArdaDateAgeFormat {
+    Coded = "coded",
+    CodedAlt = "coded-alternate",
+    Dotted = "dotted",
+    DottedAlt = "dotted-alternate",
+    Short = "short",
+    Full = "full"
+}