+++ lib/models/roleplay/job/interfaces.d.ts
@@ -1,4 +1,5 @@
-import type { Assets, JobCategoryOid, JobLevels, JobOid } from '../..';
+import type { Assets } from '../asset';
+import type { JobOid, JobCategoryOid, JobLevels } from './types';
 export interface Job {
     _id: JobOid;
     name: string;
@@ -15,4 +16,3 @@ export interface JobLevel {
     minimumIncome: Assets;
     maximumIncome: Assets;
 }
-export {};