+++ lib/models/roleplay/job/types.d.ts
@@ -1,5 +1,6 @@
-import type { JobLevel, Proficiency, ShortUuid } from '../..';
+import type { ShortUuid } from '../../common';
+import type { Proficiency } from '../meta';
+import type { JobLevel } from './interfaces';
 export type JobOid = ShortUuid;
 export type JobCategoryOid = ShortUuid;
 export type JobLevels = Record<Proficiency, JobLevel>;
-export {};