+++ lib/models/roleplay/member/types.d.ts
@@ -1,4 +1,3 @@
-import type { ShortUuid } from '../..';
+import type { ShortUuid } from '../../common';
 export type MemberOid = ShortUuid;
 export type MemberRoleOid = ShortUuid;
-export {};