+++ lib/models/roleplay/people/interfaces.d.ts
@@ -1,5 +1,6 @@
-import type { LanguageOid, PeopleGroupOid, PeopleOid } from '../..';
 import type { HexColorString } from 'discord.js';
+import type { LanguageOid } from '../language';
+import type { PeopleOid, PeopleGroupOid } from './types';
 export interface People {
     _id: PeopleOid;
     name: string;
@@ -14,7 +15,7 @@ export interface People {
 export interface PeopleGroup {
     _id: PeopleGroupOid;
     name: string;
-    height: PeopleHeight;
+    height: PeopleGroupHeight;
     /** En années ; `-1` correspond à l'immortalité. */
     lifespan: number;
 }
@@ -23,7 +24,7 @@ export interface PeopleColorScheme {
     secondary: HexColorString;
     tertiary?: HexColorString;
 }
-export interface PeopleHeight {
+export interface PeopleGroupHeight {
     /** En centimètres. */
     min: number;
     /** En centimètres. */
@@ -31,4 +32,3 @@ export interface PeopleHeight {
     /** En centimètres. */
     avg: number;
 }
-export {};