+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 
 
+## [0.2.0](https://draconocte.github.com/DracoNocte/tyalietur-lib/compare/0.1.0...0.2.0) (2023-12-04)
+
+
+### Features
+
+- **[domains]** introduce `ArdaDate`
+
+
+### Bug Fixes
+
+- **[models]** use right paths for imports
+
 ## 0.1.0 (2023-11-29)